+++ backend/JobList/Controllers/ResumeLanguagesController.cs
@@ -1,9 +1,11 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using JobList.Authorization;
 using JobList.BusinessLogic.Interfaces;
 using JobList.Common.DTOS;
 using JobList.Common.Requests;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 
@@ -13,14 +15,17 @@ namespace JobList.Controllers
     [ApiController]
     public class ResumeLanguagesController : Controller
     {
+        private readonly IAuthorizationService _authorizationService;
         private IResumeLanguagesService _resumeLanguagesService;
 
-        public ResumeLanguagesController(IResumeLanguagesService resumeLanguagesService)
+        public ResumeLanguagesController(IResumeLanguagesService resumeLanguagesService, IAuthorizationService authorizationService)
         {
+            _authorizationService = authorizationService;
             _resumeLanguagesService = resumeLanguagesService;
         }
 
         // GET: /resumeLanguages
+        [AllowAnonymous]
         [HttpGet]
         public virtual async Task<ActionResult<IEnumerable<ResumeLanguageDTO>>> Get()
         {
@@ -33,7 +38,7 @@ namespace JobList.Controllers
             return Ok(dtos);
         }
 
-
+        [AllowAnonymous]
         [HttpGet("{id}")]
         public virtual async Task<ActionResult<ResumeLanguageDTO>> GetById(int id)
         {
@@ -47,6 +52,7 @@ namespace JobList.Controllers
         }
 
         // POST: /resumeLanguages
+        [Authorize(Roles = "employee, admin")]
         [HttpPost]
         public virtual async Task<ActionResult<ResumeLanguageDTO>> Create([FromBody] ResumeLanguageRequest request)
         {
@@ -65,9 +71,18 @@ namespace JobList.Controllers
         }
 
         // PUT: /resumeLanguages/:id
+        [Authorize(Roles = "employee, admin")]
         [HttpPut("{id}")]
         public virtual async Task<ActionResult> Update([FromRoute]int id, [FromBody]ResumeLanguageRequest request)
         {
+            var isAuthorized = await _authorizationService
+                    .AuthorizeAsync(User, request.ResumeId, UserOperations.Update);
+
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -83,9 +98,20 @@ namespace JobList.Controllers
         }
 
         // DELETE: /resumeLanguages/:id
+        [Authorize(Roles = "employee, admin")]
         [HttpDelete("{id}")]
         public virtual async Task<ActionResult> Delete(int id)
         {
+            var entity = await _resumeLanguagesService.GetEntityByIdAsync(id);
+
+            var isAuthorized = await _authorizationService
+                    .AuthorizeAsync(User, entity.ResumeId, UserOperations.Delete);
+
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
+            }
+
             var result = await _resumeLanguagesService.DeleteEntityByIdAsync(id);
             if (!result)
             {
